Fix non-working split resize with mouse drag (alternative fix)
An alternative fix for the 2nd part of #1773: resize via mouse drag
doesn't work if the split contains other splits, i.e. is not a leaf
node.

Since we use GetMouseSplitID to find the node to resize, and it may
be a non-leaf node, this change makes non-leaf nodes have unique ID
just like lead nodes.