Simplify hlsearch implementation: single search state per line
To simplify the code, use a single searchState instead of a map per line.
In most scenarios it works just like with the map, however in the case
when the same file is opened in 2 split panes, with some of the lines
displayed in both panes, and with different last searches per pane, there
is a contention: each time when redrawing the screen, SearchMatch updates
the search states of those lines back and forth between these 2 searches.